update to useCallback from useEffect
